orbitto-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
bcrypt==4.0.1
PyMySQL==1.1.0
